+++ packages/insert-this-tsc-plugin/webpack.config.js
@@ -1,7 +1,10 @@
 //@ts-check
-'use strict'
 
-const path = require('path')
+'use strict';
+
+const path = require('path');
+const webpack = require('webpack');
+const CopyPlugin = require("copy-webpack-plugin");
 
 //@ts-check
 /** @typedef {import('webpack').Configuration} WebpackConfig **/
@@ -12,17 +15,13 @@ const extensionConfig = {
   mode: 'none', // this leaves the source code as close as possible to the original (when packaging we set this to 'production')
 
   entry: {
-    extension: './src/extension.ts'
+    plugin: './src/plugin.ts', 
   }, // the entry point of this extension, 📖 -> https://webpack.js.org/configuration/entry-context/
   output: {
     // the bundle is stored in the 'dist' folder (check package.json), 📖 -> https://webpack.js.org/configuration/output/
     path: path.resolve(__dirname, 'dist'),
     filename: '[name].js',
     libraryTarget: 'commonjs2'
-  },
-  externals: {
-    vscode: 'commonjs vscode' // the vscode-module is created on-the-fly and must be excluded. Add other modules that cannot be webpack'ed, 📖 -> https://webpack.js.org/configuration/externals/
-    // modules added here also need to be added in the .vscodeignore file
   },
   resolve: {
     // support reading TypeScript and JavaScript files, 📖 -> https://github.com/TypeStrong/ts-loader
@@ -43,7 +42,15 @@ const extensionConfig = {
   },
   devtool: 'nosources-source-map',
   infrastructureLogging: {
-    level: 'log' // enables logging required for problem matchers
+    level: "log", // enables logging required for problem matchers
   },
-}
-module.exports = [extensionConfig]
+  experiments: {
+    asyncWebAssembly: true,
+    // layers: true,
+    // lazyCompilation: true,
+    // outputModule: true,
+    syncWebAssembly: true,
+    // topLevelAwait: true,
+  },
+};
+module.exports = [ extensionConfig ]